NOISSUE: Renamed class and refactored the code a bit
